requirements.txt
GitPython==3.1.40
pyperclip==1.8.2
PyQt5==5.15.9